Duplicate Lock file issue

+++ django_site_queue/jsondb.py
@@ -674,19 +674,20 @@ def set_wait_queue_position(group_key):
                 for f in files:
                     if f.is_file():   
                         if str(f).endswith('.json'):
-                            try: 
-                                LOCK_PATH = str(f)+".lock" 
-                                lock = FileLock(LOCK_PATH)
-                               
-                                with lock:                                                     
-                                    sitesession = get_queue_session(f)    
-                                    sitesession['queue_position'] = position_start                                
-                                    sitesession['queue_position_epoch'] = int(time.time() * 1000)
-                                try:       
-                                    os.remove(LOCK_PATH)
-                                except Exception as k:
-                                    print ("Error Removing "+str(LOCK_PATH))
-                                    print (k)                                                                                                
+                            try:
+                                if ".lock" not in str(f):
+                                    LOCK_PATH = str(f)+".lock" 
+                                    lock = FileLock(LOCK_PATH)
+                                    
+                                    with lock:                                                     
+                                        sitesession = get_queue_session(f)    
+                                        sitesession['queue_position'] = position_start                                
+                                        sitesession['queue_position_epoch'] = int(time.time() * 1000)
+                                    try:       
+                                        os.remove(LOCK_PATH)
+                                    except Exception as k:
+                                        print ("Error Removing "+str(LOCK_PATH))
+                                        print (k)                                                                                                
                                 save_queue_session(f,sitesession)
                                 print (str(f) + "with position "+str(position_start))
                                 position_start = position_start + 1                            
@@ -721,18 +722,19 @@ def wait_queue_rotate(group_key,start, finish):
                                 
                                 
                                 try: 
-                                    LOCK_PATH = str(f)+".lock" 
-                                    lock = FileLock(LOCK_PATH)
-                                    with lock:                                    
-                                        shutil.copyfile(f, previous_path)
-                                        os.remove(f)
-                                        print ("Removing file "+str(f))
-                                    try:       
-                                        os.remove(LOCK_PATH)
-                                    except Exception as k:
-                                        print ("Error Removing "+str(LOCK_PATH))
-                                        print (k)                                        
-                                    
+                                    if ".lock" not in str(f):
+                                        LOCK_PATH = str(f)+".lock" 
+                                        lock = FileLock(LOCK_PATH)
+                                        with lock:                                    
+                                            shutil.copyfile(f, previous_path)
+                                            os.remove(f)
+                                            print ("Removing file "+str(f))
+                                        try:       
+                                            os.remove(LOCK_PATH)
+                                        except Exception as k:
+                                            print ("Error Removing "+str(LOCK_PATH))
+                                            print (k)                                        
+                                        
                                 except Exception as e:
                                     print ("Error removing "+str(f))
                                     print (e)
